requirements.txt
langgraph
langchain
langchain-community
langchain-ollama
python-dotenv
langchain_core